Move derive tests
