+++ consensus/ssz_derive/Cargo.toml
@@ -15,6 +15,3 @@ syn = "1.0.42"
 proc-macro2 = "1.0.23"
 quote = "1.0.7"
 darling = "0.13.0"
-
-[dev-dependencies]
-eth2_ssz = "0.4.0"
